Fully resolving issue #2000 and partly resolving issue #1903, A full resolution for #1903 will be implemented depending on the result of this PR.
