+++ requirements.txt
@@ -4,7 +4,7 @@ Flask-Babel==1.0.0
 Flask-Compress==1.5.0
 Flask-Login==0.5.0
 Flask-SQLAlchemy==2.4.1
-flask-restx==0.4.0
+flask-restx==0.5.1
 Jinja2==2.11.2
 SQLAlchemy==1.3.17
 Werkzeug==1.0.1
